# [IDNA] Documentation improvements, performance improvements